+++ demo/index.ts
@@ -35,7 +35,8 @@ import {DarkModeComponent} from "./components/dark-mode.component";
             path: '/about',
             component: AboutComponent,
             children: [
-                { path: '/more', component: AboutMoreComponent }
+                { path: '/more', component: AboutMoreComponent },
+                { path: '/:name', component: AboutMoreComponent }
             ]
         },
         { path: '/counter', component: CounterComponent },