+++ src/core/eventbus/v-internal-event-name.ts
@@ -2,6 +2,8 @@ export enum VInternalEventName {
     DARK_MODE_ENABLED = 'dark-mode-enabled',
     EMIT = 'emit',
     NAVIGATED = 'navigation',
+    NAVIGATION_ENDED = 'navigation_ended',
+    QUERY_PARAMS = 'query-params',
     RENDERING_STARTED = 'rendering-started',
     RENDERING_FINISHED = 'rendering-finished',
     REBUILD = 'rebuild-full-app',