+++ src/core/router/v-route-params.ts
@@ -1,3 +1,0 @@
-export interface VRouteParams {
-    [key: string]: string;
-}